# docs/swagger/cyan-business-platform.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cyan Business Platform API",
    "version": "1.0.0",
    "description": "Static OpenAPI inventory for the platform, grouped by microservice. Auth support includes Bearer JWT for endpoint/api routes and Basic Auth for internal routes."
  },
  "servers": [
    {
      "url": "{gateway_base_url}",
      "description": "Gateway base URL",
      "variables": {
        "gateway_base_url": {
          "default": "http://localhost:8080"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "paths": {
...
    "/endpoint/ai-orchestrator/drafts/{draftId}": {
      "get": {
        "tags": [
          "AI Orchestrator"
        ],
        "summary": "Get Draft",
        "description": "Get Draft for AI Orchestrator.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "{{draft_id}}"
          }
        ]
      },
      "patch": {
        "tags": [
          "AI Orchestrator"
        ],
        "summary": "Patch Draft",
        "description": "Patch Draft for AI Orchestrator.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "{{draft_id}}"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "prompt": "Spiffy-like ecommerce with CRM, notifications, and Zarinpal.",
                "title": "Spiffy Demo Updated",
                "answersPatch": {
                  "subdomain": "spiffy-shop",
                  "primaryCurrency": "IRR"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/ai-orchestrator/drafts": {
      "post": {
        "tags": [
          "AI Orchestrator Internal"
        ],
        "summary": "Create Draft Internal",
        "description": "Create Draft Internal for AI Orchestrator Internal.",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "appType": "ECOMMERCE",
                "blueprintKey": "ecommerce-crm-zarinpal-v1",
                "tenantKey": "{{tenant_key}}",
                "siteKey": "{{site_key}}",
                "clientKey": "{{client_key}}",
                "title": "Internal Spiffy Draft",
                "prompt": "Internal provisioning path.",
                "answers": {
                  "subdomain": "internal-spiffy"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "AI Orchestrator Internal"
        ],
        "summary": "List Drafts Internal",
        "description": "List Drafts Internal for AI Orchestrator Internal.",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },
    "/internal/ai-orchestrator/drafts/{draftId}": {
      "get": {
        "tags": [
          "AI Orchestrator Internal"
        ],
        "summary": "Get Draft Internal",
        "description": "Get Draft Internal for AI Orchestrator Internal.",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "{{draft_id}}"
          }
        ]
      },
      "patch": {
        "tags": [
          "AI Orchestrator Internal"
        ],
        "summary": "Patch Draft Internal",
        "description": "Patch Draft Internal for AI Orchestrator Internal.",
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "draftId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "{{draft_id}}"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "title": "Internal Updated"
              }
            }
          }
        }
      }
    },
...
  }
}